Clean up token store interfaces. - Add new methods in TokenStore interface - Remove old unused methods and constant - Remove unnecessary new interface
