--- models/classes/security/xsrf/TokenStorageInterface.php
@@ -1,60 +1,0 @@
-<?php
-
-/**
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; under version 2
- * of the License (non-upgradable).
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- *
- * Copyright (c) 2020 (original work) Open Assessment Technologies SA ;
- */
-declare(strict_types=1);
-
-namespace oat\tao\model\security\xsrf;
-
-
-interface TokenStorageInterface
-{
-    /**
-     * @param string $tokenId
-     * @return Token|null
-     */
-    public function getToken(string $tokenId): ?Token;
-
-    /**
-     * @param string $tokenId
-     * @param Token $token
-     */
-    public function setToken(string $tokenId, Token $token): void;
-
-    /**
-     * @param string $tokenId
-     * @return bool
-     */
-    public function hasToken(string $tokenId): bool;
-
-    /**
-     * @param string $tokenId
-     * @return bool
-     */
-    public function removeToken(string $tokenId): bool;
-
-    /**
-     * @return void;
-     */
-    public function clear(): void;
-
-    /**
-     * @return Token[]
-     */
-    public function getAll(): array;
-}
